+++ src/main/java/com/controlaltinsert/parkshark/employee/Employee.java
@@ -25,11 +25,11 @@ public class Employee {
         this.address = address;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
